Fixed some bugs and implemented new storage
