+++ Web/.husky/pre-commit
@@ -1,3 +1,3 @@
 #!/bin/sh
 # exit 1 will prevent commit to happen, so you can more easily debug things
-npm.cmd run prettier && npm.cmd run lint -- --fix
+
