+++ Web/package.json
@@ -26,12 +26,13 @@
     "@next-auth/prisma-adapter": "1.0.3",
     "@prisma/client": "3.15.1",
     "@react-google-maps/api": "2.12.0",
+    "aws-sdk": "2.1162.0",
     "formidable": "2.0.1",
     "framer-motion": "6.3.11",
     "moment": "2.29.3",
     "moment-timezone": "0.5.34",
     "next": "12.1.6",
-    "next-auth": "4.5.0",
+    "next-auth": "4.6.1",
     "nodemailer": "6.7.5",
     "react": "17.0.2",
     "react-dom": "17.0.2",
