+++ Web/src/helper/session.ts
@@ -1,10 +1,8 @@
-import { getSession } from 'next-auth/react';
 import { levels } from '@constants/admin';
 import { Session } from 'next-auth/core/types';
+import { getSession } from 'next-auth/react';
 
-export const currentSession = async (
-  req: any = null,
-): Promise<Session | null> => {
+export const currentSession = async (): Promise<Session | null> => {
   if (!process.env.NODE_ENV || process.env.NODE_ENV === 'development') {
     let session: Session | null = null;
     session = {
@@ -19,14 +17,7 @@ export const currentSession = async (
 
     return session;
   } else {
-    const isServer: boolean = typeof window === 'undefined';
-    let session: Session | null = null;
-    if (isServer && req !== null) {
-      session = (await getSession(req)) as Session;
-      return session;
-    } else {
-      session = (await getSession()) as Session;
-      return session;
-    }
+    let session: Session | null = (await getSession()) as Session;
+    return session;
   }
 };
