+++ Web/src/pages/api/asset/delete.ts
@@ -1,13 +1,13 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
 import { Result } from 'types/api';
 
-import { currentSession } from '@helper/session';
+import { currentSession } from '@helper/sessionServer';
 import { deleteAsset, isCreatorOfAsset } from '@helper/asset';
 import { levels } from '@constants/admin';
 import { checkerString } from '@helper/common';
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
-  const session = await currentSession(req);
+  const session = await currentSession(req, res, null, true);
 
   const { id } = req.body;
   let result: Result = {
@@ -26,7 +26,7 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
             result = {
               status: true,
               error: null,
-              msg: 'Event deleted',
+              msg: 'Asset deleted',
             };
 
             res.status(200).send(result);
