+++ Web/src/pages/api/asset/edit.ts
@@ -3,11 +3,13 @@ import { Result } from 'types/api';
 import { Asset } from 'types/asset';
 
 import { checkerString } from '@helper/common';
-import { currentSession } from '@helper/session';
+import { currentSession } from '@helper/sessionServer';
 import { editAsset } from '@helper/asset';
 import formidable, { IncomingForm } from 'formidable';
 import { promises as fs } from 'fs';
 import { levels } from '@constants/admin';
+
+import AWS from 'aws-sdk';
 
 export const config = {
   api: {
@@ -16,7 +18,7 @@ export const config = {
 };
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
-  const session = await currentSession(req);
+  const session = await currentSession(req, res, null, true);
   let result: Result = {
     status: false,
     error: '',
@@ -39,7 +41,6 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
 
       try {
         const imageFile: formidable.File = data.files.image as formidable.File;
-        let assetPath: string = '';
         let assetData: Asset | null = null;
 
         const id = data.fields.id as string;
@@ -58,25 +59,63 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
           checkerString(latitude) &&
           checkerString(longitude)
         ) {
-          if (imageFile) {
+          let success = true;
+
+          if (
+            imageFile !== null &&
+            imageFile !== undefined &&
+            process.env.S3_UPLOAD_KEY !== undefined &&
+            process.env.S3_UPLOAD_SECRET !== undefined
+          ) {
+            const s3 = new AWS.S3({
+              accessKeyId: process.env.S3_UPLOAD_KEY,
+              secretAccessKey: process.env.S3_UPLOAD_SECRET,
+            });
+
             const imagePath: string = imageFile.filepath;
-
-            assetPath = `/assets/${data.fields.eventID}_${imageFile.originalFilename}`;
-            const pathToWriteImage = `public${assetPath}`;
             const image = await fs.readFile(imagePath);
-            await fs.writeFile(pathToWriteImage, image);
 
-            assetData = {
-              id: id.trim(),
-              name: name.trim(),
-              description: description.trim(),
-              eventID: eventID.trim(),
-              visible: visible,
-              latitude: latitude.trim(),
-              longitude: longitude.trim(),
-              imagePath: assetPath.trim(),
-              createdBy: session.user.email.trim(),
-            };
+            if (process.env.S3_BUCKET_NAME !== undefined) {
+              const uploadedImage = await s3
+                .upload({
+                  Bucket: process.env.S3_BUCKET_NAME,
+                  Key: `${eventID}_${imageFile.originalFilename}`,
+                  Body: image,
+                })
+                .promise();
+
+              if (uploadedImage.Location) {
+                assetData = {
+                  id: id.trim(),
+                  name: name.trim(),
+                  description: description.trim(),
+                  eventID: eventID.trim(),
+                  visible: visible,
+                  latitude: latitude.trim(),
+                  longitude: longitude.trim(),
+                  imagePath: uploadedImage.Location.trim(),
+                };
+              } else {
+                success = false;
+                console.error('Image not uploaded');
+                result = {
+                  status: false,
+                  error: 'Image not uploaded',
+                  msg: '',
+                };
+                res.status(200).send(result);
+                res.end();
+              }
+            } else {
+              success = false;
+              result = {
+                status: false,
+                error: 'Error: no key provided for bucket',
+                msg: '',
+              };
+              res.status(200).send(result);
+              res.end();
+            }
           } else {
             assetData = {
               id: id,
@@ -86,28 +125,29 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
               visible: visible,
               latitude: latitude,
               longitude: longitude,
-              createdBy: session.user.email,
             };
           }
 
-          const createEventRequest = await editAsset(assetData);
-          if (createEventRequest.status) {
+          if (success && assetData !== null) {
+            const createEventRequest = await editAsset(assetData);
+            if (createEventRequest.status) {
+              result = {
+                status: true,
+                error: '',
+                msg: 'Successfully updated asset',
+              };
+              res.status(200).send(result);
+              res.end();
+              return;
+            }
             result = {
-              status: true,
-              error: '',
-              msg: 'Successfully updated asset',
+              status: false,
+              error: createEventRequest.error,
+              msg: '',
             };
             res.status(200).send(result);
             res.end();
-            return;
           }
-          result = {
-            status: false,
-            error: createEventRequest.error,
-            msg: '',
-          };
-          res.status(200).send(result);
-          res.end();
         } else {
           result = {
             status: false,
@@ -118,8 +158,8 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
           res.end();
         }
       } catch (error) {
-        console.log(error);
-        result = { status: false, error: 'Failed to create asset', msg: '' };
+        console.error(error);
+        result = { status: false, error: 'Failed to edit asset', msg: '' };
         res.status(200).send(result);
         res.end();
       }
