+++ Web/src/pages/api/event/create.ts
@@ -2,15 +2,14 @@ import type { NextApiRequest, NextApiResponse } from 'next';
 import { Result } from 'types/api';
 import { Event } from 'types/event';
 
-import { currentSession } from '@helper/session';
+import { currentSession } from '@helper/sessionServer';
 import { convertDateToUnix } from '@constants/date';
 import { createEvent } from '@helper/event';
 import { levels } from '@constants/admin';
 import { checkerString } from '@helper/common';
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
-  const session = await currentSession(req);
-
+  const session = await currentSession(req, res, null, true);
   const { name, description, startDate, endDate, isPublic, visible } = req.body;
   let result: Result = {
     status: false,
@@ -81,7 +80,7 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   } else {
     result = {
       status: false,
-      error: 'Unauthorized access',
+      error: 'Session not found',
       msg: null,
     };
