+++ Web/src/pages/api/event/edit.ts
@@ -2,15 +2,14 @@ import type { NextApiRequest, NextApiResponse } from 'next';
 import { Result } from 'types/api';
 import { Event } from 'types/event';
 
-import { currentSession } from '@helper/session';
+import { currentSession } from '@helper/sessionServer';
 import { convertDateToUnix } from '@constants/date';
 import { editEvent } from '@helper/event';
 import { levels } from '@constants/admin';
 import { checkerString } from '@helper/common';
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
-  const session = await currentSession(req);
-
+  const session = await currentSession(req, res, null, true);
   const { id, name, description, startDate, endDate, isPublic, visible } =
     req.body;
   let result: Result = {
@@ -39,7 +38,6 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
           endDate: end,
           isPublic: isPublic,
           visible: visible,
-          createdBy: session.user.email.trim(),
         };
 
         const event: Result = await editEvent(data);
@@ -84,7 +82,7 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   } else {
     result = {
       status: false,
-      error: 'Unauthorized access',
+      error: 'Session not found',
       msg: null,
     };
