+++ Web/src/pages/api/event/fetch.ts
@@ -3,13 +3,12 @@ import { Result } from 'types/api';
 import { Event } from 'types/event';
 
 import { formatDateToString, convertUnixToDate } from '@constants/date';
-import { currentSession } from '@helper/session';
+import { currentSession } from '@helper/sessionServer';
 import { fetchAllEventByUser } from '@helper/event';
 import { levels } from '@constants/admin';
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
-  const session = await currentSession(req);
-
+  const session = await currentSession(req, res, null, true);
   let result: Result = {
     status: false,
     error: '',
@@ -87,7 +86,7 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
       res.end();
     }
   } else {
-    result = { status: false, error: 'Unauthenticated', msg: [] };
+    result = { status: false, error: 'Session not found', msg: [] };
     res.status(200).send(result);
     res.end();
   }
