+++ Web/src/pages/api/leaderboard/fetch.ts
@@ -3,15 +3,14 @@ import { Result } from 'types/api';
 import { Event } from 'types/event';
 import { Leaderboard } from 'types/leaderboard';
 
-import { currentSession } from '@helper/session';
+import { currentSession } from '@helper/sessionServer';
 import { fetchEventByID, isEventAuthorized } from '@helper/event';
 import { fetchLeaderBoardByEventID } from '@helper/leaderboard';
 import { levels } from '@constants/admin';
 import { checkerString } from '@helper/common';
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
-  const session = await currentSession(req);
-
+  const session = await currentSession(req, res, null, true);
   let result: Result = {
     status: false,
     error: '',
@@ -86,7 +85,7 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
       res.end();
     }
   } else {
-    result = { status: false, error: 'Unauthenticated', msg: [] };
+    result = { status: false, error: 'Session not found', msg: [] };
     res.status(200).send(result);
     res.end();
   }
