+++ Web/src/pages/api/quiz/edit.ts
@@ -4,13 +4,12 @@ import { Quiz } from 'types/quiz';
 
 import { levels } from '@constants/admin';
 
-import { currentSession } from '@helper/session';
+import { currentSession } from '@helper/sessionServer';
 import { editQuiz } from '@helper/quiz';
 import { checkerNumber, checkerString } from '@helper/common';
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
-  const session = await currentSession(req);
-
+  const session = await currentSession(req, res, null, true);
   const {
     quizID,
     eventID,
@@ -56,7 +55,6 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
           answer: Number(answer),
           points: Number(points),
           visible: visible,
-          createdBy: session.user.email.trim(),
         };
 
         const qn = await editQuiz(data);
@@ -103,7 +101,7 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   } else {
     result = {
       status: false,
-      error: 'Unauthorized access',
+      error: 'Session not found',
       msg: null,
     };
