+++ Web/src/pages/api/unity/event.ts
@@ -77,20 +77,12 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
         res.end();
       }
     } else {
-      result = {
-        status: false,
-        error: 'Unauthorized',
-        msg: [],
-      };
+      result = { status: false, error: 'Unauthorized, invalid token', msg: [] };
       res.status(200).send(result);
       res.end();
     }
   } else {
-    result = {
-      status: false,
-      error: 'Unauthorized',
-      msg: [],
-    };
+    result = { status: false, error: 'Unauthorized, token not found', msg: [] };
     res.status(200).send(result);
     res.end();
   }
