+++ Web/src/pages/api/unity/join.ts
@@ -52,8 +52,8 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
       } else {
         result = {
           status: false,
-          error: 'Unauthorized',
-          msg: '',
+          error: 'Unauthorized, invalid token',
+          msg: [],
         };
         res.status(200).send(result);
         res.end();
@@ -61,8 +61,8 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
     } else {
       result = {
         status: false,
-        error: 'Unauthorized',
-        msg: '',
+        error: 'Unauthorized, token not found',
+        msg: [],
       };
       res.status(200).send(result);
       res.end();
