+++ Web/src/pages/api/unity/points.ts
@@ -53,12 +53,20 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
           res.end();
         }
       } else {
-        result = { status: false, error: 'Unauthorized', msg: '' };
+        result = {
+          status: false,
+          error: 'Unauthorized, invalid token',
+          msg: [],
+        };
         res.status(200).send(result);
         res.end();
       }
     } else {
-      result = { status: false, error: 'Unauthorized', msg: '' };
+      result = {
+        status: false,
+        error: 'Unauthorized, token not found',
+        msg: [],
+      };
       res.status(200).send(result);
       res.end();
     }
